requirements.txt
# --- Web & API ---
fastapi==0.111.0
uvicorn==0.23.2
requests==2.31.0
pydantic==2.5.3
python-dotenv==1.0.0
streamlit==1.37.0

# --- AI / NLP ---
google-generativeai==0.6.0
langchain==0.1.17
langchain_community
sentence-transformers==2.2.2

# --- Vector DB ---
pymilvus==2.4.1
numpy==1.26.4

# --- Infra ---
redis==5.0.3
websocket-client==1.8.0
